client/server common code  extraction
