+++ src/main/scala/io/quartz/http2/Http2Connection.scala
@@ -30,7 +30,7 @@ import io.quartz.http2.routes.HttpRoute
 
 object Http2Connection {
 
-  val STREAMTBL_PURGE_DELAY = 0 // 320 last streams will be recognozable while in closed state
+  val STREAMTBL_PURGE_DELAY = 300 // 320 last streams will be recognozable while in closed state
 
   private[this] def outBoundWorkerProc(
       ch: IOChannel,
@@ -106,7 +106,7 @@ object Http2Connection {
       streamId: Int,
       received: Ref[Long],
       window: Ref[Long],
-      len : Int
+      len: Int
   ) =
     for {
       bytes_received <- received.getAndUpdate(_ - len)
@@ -122,7 +122,7 @@ object Http2Connection {
       )
     } yield (send_update)
 
-  private[this] def parseFrame(bb: ByteBuffer) = {
+  def parseFrame(bb: ByteBuffer) = {
     val sbb = bb.slice();
 
     val len = Frames.getLengthField(sbb)
@@ -147,8 +147,8 @@ object Http2Connection {
       _ <- ZIO.when(o_stream.isEmpty)(ZIO.fail(ErrorGen(streamId, Error.FRAME_SIZE_ERROR, "invalid stream id")))
       stream: Http2StreamCommon <- ZIO.attempt(o_stream.get)
       //////////////////////////////////////////////////////////////////////////////////////////////////////////
-      _ <- windowsUpdate[Env](c, 0, c.globalBytesOfPendingInboundData, c.globalInboundWindow, len )
-      _ <- windowsUpdate[Env](c, streamId, stream.bytesOfPendingInboundData, stream.inboundWindow, len )
+      _ <- windowsUpdate[Env](c, 0, c.globalBytesOfPendingInboundData, c.globalInboundWindow, len)
+      _ <- windowsUpdate[Env](c, streamId, stream.bytesOfPendingInboundData, stream.inboundWindow, len)
     } yield (bb)
   }
 
@@ -226,7 +226,12 @@ object Http2Connection {
   }
 }
 
-trait Http2StreamCommon(val bytesOfPendingInboundData: Ref[Long], val inboundWindow: Ref[Long])
+trait Http2StreamCommon(
+    val bytesOfPendingInboundData: Ref[Long],
+    val inboundWindow: Ref[Long],
+    val transmitWindow: Ref[Long],
+    val outXFlowSync: Queue[Unit]
+)
 
 class Http2Stream(
     active: Ref[Boolean],
@@ -234,29 +239,19 @@ class Http2Stream(
     val header: ArrayBuffer[ByteBuffer],
     val trailing_header: ArrayBuffer[ByteBuffer],
     val inDataQ: Queue[ByteBuffer], // accumulate data packets in stream function
-    val outXFlowSync: Queue[Unit], // flow control sync queue for data frames
-    val transmitWindow: Ref[Long],
+    outXFlowSync: Queue[Unit], // flow control sync queue for data frames
+    transmitWindow: Ref[Long],
     syncUpdateWindowQ: Queue[Unit],
     bytesOfPendingInboundData: Ref[Long],
     inboundWindow: Ref[Long],
     val contentLenFromHeader: Promise[Throwable, Option[Int]],
     val trailingHeader: Promise[Throwable, Headers],
     val done: Promise[Throwable, Unit]
-) extends Http2StreamCommon(bytesOfPendingInboundData, inboundWindow) {
+) extends Http2StreamCommon(bytesOfPendingInboundData, inboundWindow, transmitWindow, outXFlowSync) {
   var endFlag = false // half-closed if true
   var endHeadersFlag = false
   var contentLenFromDataFrames = 0
 
-}
-
-trait Http2ConnectionCommon(
-    val INITIAL_WINDOW_SIZE: Int,
-    val globalBytesOfPendingInboundData: Ref[Long],
-    val globalInboundWindow: Ref[Long],
-    val hSem2: Semaphore
-) {
-  def sendFrame(b: ByteBuffer): zio.UIO[Boolean]
-  def getStream(id: Int): Option[Http2StreamCommon]
 }
 
 class Http2Connection[Env](
@@ -274,7 +269,14 @@ class Http2Connection[Env](
     MAX_CONCURRENT_STREAMS: Int,
     HTTP2_KEEP_ALIVE_MS: Int,
     INITIAL_WINDOW_SIZE: Int
-) extends Http2ConnectionCommon(INITIAL_WINDOW_SIZE, globalBytesOfPendingInboundData, globalInboundWindow, hSem2) {
+) extends Http2ConnectionCommon(
+      INITIAL_WINDOW_SIZE,
+      globalBytesOfPendingInboundData,
+      globalInboundWindow,
+      globalTransmitWindow,
+      outq,
+      hSem2
+    ) {
 
   val settings: Http2Settings = new Http2Settings()
   val settings_client = new Http2Settings()
@@ -693,81 +695,9 @@ class Http2Connection[Env](
     } yield (b)
   }
 
-  case class txWindow_SplitDataFrame(buffer: ByteBuffer, dataLen: Int)
 
-  private[this] def splitDataFrames(
-      bb: ByteBuffer,
-      requiredLen: Long
-  ): (txWindow_SplitDataFrame, Option[txWindow_SplitDataFrame]) = {
-    val original_bb = bb.slice()
-    val len = Frames.getLengthField(bb)
-    val frameType = bb.get()
-    val flags = bb.get()
-    val streamId = Frames.getStreamId(bb)
-
-    val endStream = if ((flags & Flags.END_STREAM) != 0) true else false
-
-    if (requiredLen < len) {
-      val buf0 = Array.ofDim[Byte](requiredLen.toInt)
-      bb.get(buf0)
-
-      val dataFrame1 = Frames.mkDataFrame(streamId, false, padding = 0, ByteBuffer.wrap(buf0))
-      val dataFrame2 = Frames.mkDataFrame(streamId, endStream, padding = 0, bb)
-
-      (
-        txWindow_SplitDataFrame(dataFrame1, requiredLen.toInt),
-        Some(txWindow_SplitDataFrame(dataFrame2, len - requiredLen.toInt))
-      )
-
-    } else (txWindow_SplitDataFrame(original_bb, len), None)
-  }
-
-  private[this] def txWindow_Transmit(stream: Http2Stream, bb: ByteBuffer, data_len: Int): Task[Long] = {
-    for {
-      tx_g <- globalTransmitWindow.get
-      tx_l <- stream.transmitWindow.get
-      bytesCredit <- ZIO.succeed(Math.min(tx_g, tx_l))
-
-      _ <-
-        if (bytesCredit > 0)
-          (for {
-            rlen <- ZIO.succeed(Math.min(bytesCredit, data_len))
-            frames <- ZIO.attempt(splitDataFrames(bb, rlen))
-            _ <- sendFrame(frames._1.buffer)
-            _ <- globalTransmitWindow.update(_ - rlen)
-            _ <- stream.transmitWindow.update(_ - rlen)
-
-            _ <- frames._2 match {
-              case Some(f0) =>
-                stream.outXFlowSync.take *> txWindow_Transmit(stream, f0.buffer, f0.dataLen)
-              case None => ZIO.unit
-            }
-
-          } yield ())
-        else stream.outXFlowSync.take *> txWindow_Transmit(stream, bb, data_len)
-
-    } yield (bytesCredit)
-  }
-
-  def sendDataFrame(streamId: Int, bb: => ByteBuffer): Task[Unit] =
-    for {
-      t <- ZIO.attempt(Http2Connection.parseFrame(bb))
-      len = t._1
-      _ <- ZIO.logTrace(s"sendDataFrame() - $len bytes")
-      opt_D <- ZIO.attempt(streamTbl.get(streamId))
-      _ <- opt_D match {
-        case Some(ce) =>
-          for {
-            _ <- txWindow_Transmit(ce, bb, len)
-          } yield ()
-        case None => ZIO.logError("sendDataFrame lost streamId")
-      }
-    } yield ()
-
-  def sendFrame(b: ByteBuffer) = outq.offer(b)
 
   ////////////////////////////////////////////
-
   def route2(streamId: Int, request: Request): ZIO[Env, Throwable, Unit] = {
 
     val T = for {
@@ -816,7 +746,14 @@ class Http2Connection[Env](
 
             _ <- ZIO.scoped {
               hSem.withPermitScoped *> ZIO.foreach(
-                headerFrame(streamId, Priority.NoPriority, endStreamInHeaders, response.headers)
+                headerFrame(
+                  streamId,
+                  settings,
+                  Priority.NoPriority,
+                  endStreamInHeaders,
+                  headerEncoder,
+                  response.headers
+                )
               )((b => sendFrame(b)))
             }
             /*
@@ -834,7 +771,7 @@ class Http2Connection[Env](
                     for {
                       chunk0 <- pref.get
                       _ <- ZIO.when(chunk0.nonEmpty)(
-                        ZIO.foreach(dataFrame(streamId, false, ByteBuffer.wrap(chunk0.toArray)))(b =>
+                        ZIO.foreach(dataFrame(settings, streamId, false, ByteBuffer.wrap(chunk0.toArray)))(b =>
                           sendDataFrame(streamId, b)
                         )
                       )
@@ -847,7 +784,7 @@ class Http2Connection[Env](
             lastChunk <- pref.get
             _ <- (ZIO
               .when(endStreamInHeaders == false)(
-                ZIO.foreach(dataFrame(streamId, true, ByteBuffer.wrap(lastChunk.toArray)))(b =>
+                ZIO.foreach(dataFrame(settings,streamId, true, ByteBuffer.wrap(lastChunk.toArray)))(b =>
                   sendDataFrame(streamId, b)
                 )
               ))
@@ -868,7 +805,9 @@ class Http2Connection[Env](
             _ <- ZIO.scoped {
               hSem.withPermitScoped.tap(_ =>
                 for {
-                  bb2 <- ZIO.succeed(headerFrame(streamId, Priority.NoPriority, true, o44.headers))
+                  bb2 <- ZIO.succeed(
+                    headerFrame(streamId, settings, Priority.NoPriority, true, headerEncoder, o44.headers)
+                  )
                   _ <- ZIO.foreach(bb2)(b => sendFrame(b))
                   _ <- updateStreamWith(10, streamId, c => c.done.succeed(()).unit)
 
@@ -890,93 +829,6 @@ class Http2Connection[Env](
       _ <- ZIO.logDebug(s"Close stream: $streamId")
     } yield ()
 
-  }
-
-  /** Generate stream data frame(s) for the specified data
-    *
-    * If the data exceeds the peers MAX_FRAME_SIZE setting, it is fragmented into a series of frames.
-    */
-  def dataFrame(
-      streamId: Int,
-      endStream: Boolean,
-      data: ByteBuffer
-  ): scala.collection.immutable.Seq[ByteBuffer] = {
-    val limit =
-      settings.MAX_FRAME_SIZE - 128
-
-    if (data.remaining <= limit) {
-
-      val acc = new ArrayBuffer[ByteBuffer]
-      acc.addOne(Frames.mkDataFrame(streamId, endStream, padding = 0, data))
-
-      acc.toSeq
-    } else { // need to fragment
-      // println("FRAGMENT - SHOULD NOT BE THERE")
-      val acc = new ArrayBuffer[ByteBuffer]
-
-      while (data.hasRemaining) {
-
-        val len = math.min(data.remaining(), limit)
-
-        val cur_pos = data.position()
-
-        val thisData = data.slice.limit(len)
-
-        data.position(cur_pos + len)
-
-        val eos = endStream && !data.hasRemaining
-        acc.addOne(Frames.mkDataFrame(streamId, eos, padding = 0, thisData))
-      }
-      // acc.foreach(c => println("------>" + c.limit()))
-      acc.toSeq
-    }
-  }
-
-  def takeSlice(buf: ByteBuffer, len: Int): ByteBuffer = {
-    val head = buf.slice.limit(len)
-    buf.position(len)
-    head
-  }
-
-  /** Generate stream header frames from the provided header sequence
-    *
-    * If the compressed representation of the headers exceeds the MAX_FRAME_SIZE setting of the peer, it will be broken
-    * into a HEADERS frame and a series of CONTINUATION frames.
-    */
-  //////////////////////////////
-  def headerFrame(
-      streamId: Int,
-      priority: Priority,
-      endStream: Boolean,
-      headers: Headers
-  ): scala.collection.immutable.Seq[ByteBuffer] = {
-    val rawHeaders = headerEncoder.encodeHeaders(headers)
-
-    val limit = settings.MAX_FRAME_SIZE - 61
-
-    val headersPrioritySize =
-      if (priority.isDefined) 5 else 0 // priority(4) + weight(1), padding = 0
-
-    if (rawHeaders.remaining() + headersPrioritySize <= limit) {
-      val acc = new ArrayBuffer[ByteBuffer]
-      acc.addOne(Frames.mkHeaderFrame(streamId, priority, endHeaders = true, endStream, padding = 0, rawHeaders))
-
-      acc.toSeq
-    } else {
-      // need to fragment
-      val acc = new ArrayBuffer[ByteBuffer]
-
-      val headersBuf = takeSlice(rawHeaders, limit - headersPrioritySize)
-      acc += Frames.mkHeaderFrame(streamId, priority, endHeaders = false, endStream, padding = 0, headersBuf)
-
-      while (rawHeaders.hasRemaining) {
-        val size = math.min(limit, rawHeaders.remaining)
-        val continueBuf = takeSlice(rawHeaders, size)
-        val endHeaders = !rawHeaders.hasRemaining
-        acc += Frames.mkContinuationFrame(streamId, endHeaders, continueBuf)
-      }
-      acc.toSeq
-    }
   }
 
   def processIncoming(leftOver: Chunk[Byte]): ZIO[Env, Throwable, Unit] = {
